+++ api/layer/oracle/tx_grpc.pb.go
@@ -24,8 +24,6 @@ const (
 	Msg_Tip_FullMethodName                  = "/layer.oracle.Msg/Tip"
 	Msg_UpdateCyclelist_FullMethodName      = "/layer.oracle.Msg/UpdateCyclelist"
 	Msg_UpdateQueryDataLimit_FullMethodName = "/layer.oracle.Msg/UpdateQueryDataLimit"
-	Msg_BatchSubmitValue_FullMethodName     = "/layer.oracle.Msg/BatchSubmitValue"
-	Msg_UpdateMaxBatchSize_FullMethodName   = "/layer.oracle.Msg/UpdateMaxBatchSize"
 )
 
 // MsgClient is the client API for Msg service.
@@ -41,8 +39,6 @@ type MsgClient interface {
 	Tip(ctx context.Context, in *MsgTip, opts ...grpc.CallOption) (*MsgTipResponse, error)
 	UpdateCyclelist(ctx context.Context, in *MsgUpdateCyclelist, opts ...grpc.CallOption) (*MsgUpdateCyclelistResponse, error)
 	UpdateQueryDataLimit(ctx context.Context, in *MsgUpdateQueryDataLimit, opts ...grpc.CallOption) (*MsgUpdateQueryDataLimitResponse, error)
-	BatchSubmitValue(ctx context.Context, in *MsgBatchSubmitValue, opts ...grpc.CallOption) (*MsgBatchSubmitValueResponse, error)
-	UpdateMaxBatchSize(ctx context.Context, in *MsgUpdateMaxBatchSize, opts ...grpc.CallOption) (*MsgUpdateMaxBatchSizeResponse, error)
 }
 
 type msgClient struct {
@@ -103,26 +99,6 @@ func (c *msgClient) UpdateQueryDataLimit(ctx context.Context, in *MsgUpdateQuery
 	return out, nil
 }
 
-func (c *msgClient) BatchSubmitValue(ctx context.Context, in *MsgBatchSubmitValue, opts ...grpc.CallOption) (*MsgBatchSubmitValueResponse, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
-	out := new(MsgBatchSubmitValueResponse)
-	err := c.cc.Invoke(ctx, Msg_BatchSubmitValue_FullMethodName, in, out, cOpts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *msgClient) UpdateMaxBatchSize(ctx context.Context, in *MsgUpdateMaxBatchSize, opts ...grpc.CallOption) (*MsgUpdateMaxBatchSizeResponse, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
-	out := new(MsgUpdateMaxBatchSizeResponse)
-	err := c.cc.Invoke(ctx, Msg_UpdateMaxBatchSize_FullMethodName, in, out, cOpts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
 // MsgServer is the server API for Msg service.
 // All implementations must embed UnimplementedMsgServer
 // for forward compatibility.
@@ -136,8 +112,6 @@ type MsgServer interface {
 	Tip(context.Context, *MsgTip) (*MsgTipResponse, error)
 	UpdateCyclelist(context.Context, *MsgUpdateCyclelist) (*MsgUpdateCyclelistResponse, error)
 	UpdateQueryDataLimit(context.Context, *MsgUpdateQueryDataLimit) (*MsgUpdateQueryDataLimitResponse, error)
-	BatchSubmitValue(context.Context, *MsgBatchSubmitValue) (*MsgBatchSubmitValueResponse, error)
-	UpdateMaxBatchSize(context.Context, *MsgUpdateMaxBatchSize) (*MsgUpdateMaxBatchSizeResponse, error)
 	mustEmbedUnimplementedMsgServer()
 }
 
@@ -162,12 +136,6 @@ func (UnimplementedMsgServer) UpdateCyclelist(context.Context, *MsgUpdateCycleli
 }
 func (UnimplementedMsgServer) UpdateQueryDataLimit(context.Context, *MsgUpdateQueryDataLimit) (*MsgUpdateQueryDataLimitResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateQueryDataLimit not implemented")
-}
-func (UnimplementedMsgServer) BatchSubmitValue(context.Context, *MsgBatchSubmitValue) (*MsgBatchSubmitValueResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method BatchSubmitValue not implemented")
-}
-func (UnimplementedMsgServer) UpdateMaxBatchSize(context.Context, *MsgUpdateMaxBatchSize) (*MsgUpdateMaxBatchSizeResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method UpdateMaxBatchSize not implemented")
 }
 func (UnimplementedMsgServer) mustEmbedUnimplementedMsgServer() {}
 func (UnimplementedMsgServer) testEmbeddedByValue()             {}
@@ -280,42 +248,6 @@ func _Msg_UpdateQueryDataLimit_Handler(srv interface{}, ctx context.Context, dec
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Msg_BatchSubmitValue_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(MsgBatchSubmitValue)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(MsgServer).BatchSubmitValue(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: Msg_BatchSubmitValue_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(MsgServer).BatchSubmitValue(ctx, req.(*MsgBatchSubmitValue))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Msg_UpdateMaxBatchSize_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(MsgUpdateMaxBatchSize)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(MsgServer).UpdateMaxBatchSize(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: Msg_UpdateMaxBatchSize_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(MsgServer).UpdateMaxBatchSize(ctx, req.(*MsgUpdateMaxBatchSize))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 // Msg_ServiceDesc is the grpc.ServiceDesc for Msg service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -342,14 +274,6 @@ var Msg_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "UpdateQueryDataLimit",
 			Handler:    _Msg_UpdateQueryDataLimit_Handler,
-		},
-		{
-			MethodName: "BatchSubmitValue",
-			Handler:    _Msg_BatchSubmitValue_Handler,
-		},
-		{
-			MethodName: "UpdateMaxBatchSize",
-			Handler:    _Msg_UpdateMaxBatchSize_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},